ref: move submetrics measure to criterion bench in plonky2
